Fix missing endpoint for self hosted
