+++ apps/studio/data/config/project-settings-v2-query.ts
@@ -22,6 +22,7 @@ export async function getProjectSettings(
   })
 
   if (error) handleError(error)
+  console.log(data)
   return data as unknown as ProjectSettings
 }
